Removes validations from LA address county
The ignore_validation key is included with the county sub_fields when
building the address question. This attribute is used to define the
required_sub_fields for the question. The address_question_html then
checks if the county question is required and adds a class depending on
the result.

Select fields are added to the validateSingleQuestion function to avoid
validation when the 'not-required' class is present.